requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
chromadb
unstructured
pypdf
pillow

